Add MariaDataFileLoader
This commit adds the starting implementation of the MaraiDataFileLoader
class. It also adds some interfaces that go along with it. The
MariaDatFileLoader will be a loader that specializes in transforming
raw *.data files generated by our MariaDB export process and it will
turn them into standardized mapped arrays for use by transformation
code.